clean up the MultiSig block height logic
as to make it less technical, responding to the code review
